fix(crypto): apply internal pre-audit hardening (F-2/F-4/F-5/F-9/F-10)
Confirmed findings from the 2026-07-15 internal crypto pre-audit; testnet
tightening ahead of the mandatory external audit (no live mainnet activation).

- F-4: clsag::verify rejects an identity key image (mirrors dual_clsag).
- F-5: range-proof bit_challenge encodes the bit index as fixed-width u32 (was
  platform-width usize) — removes a 32/64-bit consensus divergence in a
  consensus-critical hash preimage. (No on-chain range proofs exist to migrate:
  live confidential-output count is 0.)
- F-9: reject a degenerate identity ephemeral R at consensus (single choke point
  parse_confidential_output, covering both shield + confidential output loops)
  and in the recipient scan path — blocks a sender-chosen deanonymization.
- F-10: apply_block_write_with_commitment now refuses confidential/shield txs
  instead of silently materializing them as transparent UTXOs (a latent reorg
  landmine on the non-live path); the live path is apply_block_dag_ordered.
- F-2: correct ShadowHash's false 'memory-hard / ASIC-resistant' claims — its
  scratchpad access pattern is not value-dependent, so it is streamable and
  reduces to SHA256 work; documented as legacy pre-fork only.

F-1 (single UmbraHash cost floor / retire ShadowHash) is a fork-activation +
governance decision and an external-auditor design question — NOT changed here.
Full suite green (2319 lib + targeted crypto tests).

+++ engine/privacy/confidential/range_proof.rs
@@ -212,7 +212,11 @@ pub fn verify(commitment: &RistrettoPoint, proof: &RangeProof) -> bool {
 fn bit_challenge(bit_index: usize, point: &RistrettoPoint) -> Scalar {
     let mut hasher = Sha512::new();
     hasher.update(b"ShadowDAG_range_v1");
-    hasher.update(bit_index.to_le_bytes());
+    // Fixed-width encoding: `usize::to_le_bytes()` is platform-width (8 bytes on
+    // 64-bit, 4 on 32-bit), which would make the challenge — and thus the proof
+    // bytes and their verification — differ by architecture, a cross-arch
+    // consensus divergence. Indices are 0..63, so u32 is always sufficient.
+    hasher.update((bit_index as u32).to_le_bytes());
     hasher.update(point.compress().as_bytes());
     Scalar::from_hash(hasher)
 }
